refactor(routes): move thumbnails to asset router
