+++ docs/ARCHITECTURE_PHASE_1_ROADMAP.md
@@ -127,6 +127,7 @@ Progress:
 - Extracted brand/static/PWA endpoints and static asset helpers into
   `app/relaytv_app/routes/assets.py`, while keeping QR connect rendering in
   the aggregate router for now.
+- Extracted `/thumbs/{filename}` into `app/relaytv_app/routes/assets.py`.
 
 ### M3: Extract Queue And Playback Routers
 
@@ -255,6 +256,7 @@ Add entries here as PRs land into `codex/architecture-phase-1`.
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting `/discovery/status` and `/tv/status`. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with assets or app info. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting `/app/info` and release update-check helpers. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with static/PWA/assets routes or simple root redirect. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by extracting brand/static/PWA asset routes and shared static asset helpers. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with thumbnails/snapshots/uploads or simple root redirect. |
+| 2026-06-30 | local | `codex/architecture-phase-1` | Continued M2 by moving thumbnail serving into the asset router. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M2 with snapshots/uploads or simple root redirect. |
 
 ## Open Questions
 
@@ -264,6 +266,6 @@ Add entries here as PRs land into `codex/architecture-phase-1`.
 
 ## Current Recommendation
 
-Continue M2 with low-risk thumbnails, snapshots, uploads, or simple root
-redirect extraction. Avoid moving playback, settings, and Jellyfin routes until
-the router aggregation pattern is proven.
+Continue M2 with low-risk snapshots, uploads, or simple root redirect
+extraction. Avoid moving playback, settings, and Jellyfin routes until the
+router aggregation pattern is proven.
